added acc info graphic

+++ src/Library/CrossCosinuss/Device/Sensors/Accelerometer.cs
@@ -2,13 +2,13 @@
 {
     public class Accelerometer
     {
-        public int X { get; private set; } = 0;
+        public float X { get; private set; } = 0;
 
-        public int Y { get; private set; } = 0;
+        public float Y { get; private set; } = 0;
 
-        public int Z { get; private set; } = 0;
+        public float Z { get; private set; } = 0;
 
-        public Accelerometer(int x, int y, int z)
+        public Accelerometer(float x, float y, float z)
         {
             this.X = x;
             this.Y = y;
